+++ docs/components/content/examples/vue/separator/ExampleVueSeparatorBasic.vue
@@ -1,0 +1,3 @@
+<template>
+  <NSeparator label="OR" class="mx-2" />
+</template>