+++ packages/nuxt/playground/app.vue
@@ -1,3 +1,6 @@
+<script setup lang="ts">
+</script>
+
 <template>
   <div class="h-screen w-screen flex flex-col items-center gap-4 bg-base p-16">
     <div class="flex space-x-2">