+++ packages/nuxt/src/runtime/components/forms/select/SelectTrigger.vue
@@ -22,7 +22,7 @@ const statusClassVariants = computed(() => {
     success: 'btn-outline-success',
     warning: 'btn-outline-warning',
     error: 'btn-outline-error',
-    default: props?.una?.btnDefaultVariant ?? 'btn-outline-white',
+    default: props.select ? `select-${props.select}` : 'select-default-variant',
   }
 
   const text = {
@@ -30,7 +30,7 @@ const statusClassVariants = computed(() => {
     success: 'text-success',
     warning: 'text-warning',
     error: 'text-error',
-    default: 'text-muted',
+    default: '',
   }
 
   const icon = {