+++ packages/nuxt/src/runtime/types/index.ts
@@ -18,7 +18,9 @@ export * from './checkbox'
 export * from './slider'
 export * from './progress'
 export * from './skeleton'
+export * from './tabs'
 export * from './select'
+export * from './separator'
 
 export interface Colors {
   [key: string]: string