+++ uno.config.ts
@@ -38,6 +38,8 @@ export default defineConfig({
   configDeps: [
     '../packages/nuxt/src/preset.config.ts',
 
+    '../packages/preset/src/_shortcuts/tabs.ts',
+    '../packages/preset/src/_shortcuts/separator.ts',
     '../packages/preset/src/_shortcuts/select.ts',
     '../packages/preset/src/_shortcuts/skeleton.ts',
     '../packages/preset/src/_shortcuts/progress.ts',